Rename file, use type guards
