+++ src/index.ts
@@ -1,7 +1,7 @@
 export { WindowPostMessageStream } from './WindowPostMessageStream';
 export { WorkerPostMessageStream } from './WorkerPostMessageStream';
 export { WorkerParentPostMessageStream } from './WorkerParentPostMessageStream';
-export { ChildProcessParentMessageStream } from './ChildProcessParentMessageStream';
+export { ParentProcessMessageStream } from './ParentProcessMessageStream';
 export { ChildProcessMessageStream } from './ChildProcessMessageStream';
 export { PostMessageEvent } from './BasePostMessageStream';
 export { StreamData, StreamMessage } from './utils';
